+++ src/main/java/br/com/powtec/finance/database/library/model/dto/CreditCardInstallmentDTO.java
@@ -1,5 +1,6 @@
 package br.com.powtec.finance.database.library.model.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import br.com.powtec.finance.database.library.enums.EntryTypeEnum;
@@ -8,7 +9,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
 @AllArgsConstructor
 @Builder
 @Getter
@@ -19,7 +19,7 @@ public class CreditCardInstallmentDTO {
   private Long id;
   private Integer installment;
   private String referenceMonth;
-  private Double value;
+  private BigDecimal value;
   private EntryTypeEnum entryType;
   private CreditCardMovementDTO movement;
   private CreditCardStatementDTO statement;
